reuse tokenmiddleware in webauthhandler
concourse/concourse#3141
